move simplifyfoods into its own helper file to prevent recursive dependencies
