+++ src/lib/metabolism.ts
@@ -1,5 +1,4 @@
 import { InsulinVariantManager } from "../managers/insulinVariantManager";
-import type Food from "../models/food";
 import type { InsulinVariant } from "../models/types/insulinVariant";
 import type { RescueVariant } from "../models/types/rescueVariant";
 import { CalibrationStore } from "../storage/calibrationStore";
@@ -97,22 +96,4 @@ export function getApproximatedProfile() {
     carbsEffect: alphaCarbs,
     proteinEffect: alphaProtein,
   };
-}
-export function simplifyFoods(foods: Food[]) {
-  let simplified: Food[] = [];
-  for (let food of foods) {
-    let existingFood: Food | null = null;
-    for (let f of simplified) {
-      if (food.name === f.name) {
-        existingFood = f;
-        break;
-      }
-    }
-    if (!existingFood) {
-      simplified.push(food);
-      continue;
-    }
-    existingFood.amount += food.amount;
-  }
-  return simplified;
 }
